add Car Controller and Command / Results

+++ BE.DAL/Client.cs
@@ -1,10 +1,8 @@
-using System;
-
 namespace BE.DAL
 {
     public class Client
     {
-        public int ClientId { get; set; }
+        public long ClientId { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string PhoneNumber { get; set; }
